[mips] Fix incorrect cast
Fixes the following compilation error when building with GCC:
error: call of overloaded ‘Operand(int64_t)’ is ambiguous

+++ src/compiler/backend/mips/code-generator-mips.cc
@@ -4089,7 +4089,7 @@ void CodeGenerator::AssembleReturn(InstructionOperand* additional_pop_count) {
     } else if (FLAG_debug_code) {
       __ Assert(eq, AbortReason::kUnexpectedAdditionalPopValue,
                 g.ToRegister(additional_pop_count),
-                Operand(static_cast<int64_t>(0)));
+                Operand(static_cast<int32_t>(0)));
     }
   }
   // Functions with JS linkage have at least one parameter (the receiver).
